pyproject.toml
[tool.poetry]
name = "hugging4all"
version = "0.1.0"
description = ""
authors = ["Omar U. Espejel <espejelomar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.30.0"
pyyaml = "^6.0"
termcolor = "^2.3.0"
tqdm = "^4.65.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
